Split some code to new files
Move the derivation ATerm code to new files
-------------------------------------------

Like the earlier move of the JSON code to `derivation-json.cc`: the
single `derivations.{hh,cc}` was getting too big.

This moves the ATerm parser (`parseDerivation` and friends), the ATerm
printer (`unparse` and friends), the hash-modulo machinery (which
operates on ATerm-family intermediate representations), and the
wire-protocol (de)serialization of `BasicDerivation` (which also uses
the legacy ATerm-style representation) to a new
`derivation/aterm.{hh,cc}` pair, with the header taking over the
corresponding declarations from `derivations.hh`.

Almost nothing is changed, this is nearly a pure move. The exceptions:

* `hashDerivationModulo` (the output-masking variant, used by
  `processDerivationOutputPaths` which stays in `derivations.cc`) is no
  longer `static`, and now declared in the new header.

* Users of the moved declarations gain an include of the new header.

Split out "derivation/elaboration" from "derivation-options"
------------------------------------------------------------

Once Derivation owns the options in the next commit, the "elaboration"
code here will only be relevant for converting between the full-featured
in-memory derivation and the on-disk ATerm format. This split separates
it out ahead of that change.

Split out `derivation/output.{hh,cc}` from `derivations.{hh,cc}`
----------------------------------------------------------------
`DerivationOutput` (and the related maps/aliases) get their own header
and source file, `derivation/output.{hh,cc}`. This is another pure move
to make `derivations.{hh,cc}` smaller; the next commit will grow the
definitions around the outputs map, so having the output type separate
first keeps that diff focused.
