+++ src/libstore/daemon.cc
@@ -16,6 +16,7 @@
 #include "nix/store/path-with-outputs.hh"
 #include "nix/util/finally.hh"
 #include "nix/util/archive.hh"
+#include "nix/store/derivation/aterm.hh"
 #include "nix/store/derivations.hh"
 #include "nix/util/args.hh"
 #include "nix/util/logging.hh"
