+++ src/libstore/include/nix/store/meson.build
@@ -36,6 +36,10 @@ headers = [ config_pub_h ] + files(
   'derivation-options.hh',
   'derivation/full-inputs.hh',
   'derivations.hh',
+  'derivation/aterm.hh',
+  'derivation/elaborate.hh',
+  'derivation/full-inputs.hh',
+  'derivation/output.hh',
   'derived-path-map.hh',
   'derived-path.hh',
   'downstream-placeholder.hh',
