+++ src/libstore/meson.build
@@ -278,9 +278,14 @@ sources = files(
   'common-ssh-store-config.cc',
   'content-address.cc',
   'daemon.cc',
+  'derivation/aterm.cc',
+  'derivation/elaborate.cc',
   'derivation-json.cc',
   'derivation-options.cc',
+  'derivation/aterm.cc',
+  'derivation/elaborate.cc',
   'derivation/full-inputs.cc',
+  'derivation/output.cc',
   'derivations.cc',
   'derived-path-map.cc',
   'derived-path.cc',
